requirements.txt
pandas>=2.0.0
scikit-learn>=1.3.0
psycopg2-binary>=2.9.0
Flask>=3.0.0
Flask-CORS>=4.0.0
python-dotenv>=1.0.0